requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
openai
pydantic